reorganized into proper search class a la dds and aaa
