+++ README.md
@@ -1,4 +1,4 @@
-EODMS API Client
+eodms-py EODMS Client Package
 ================
 
 The EODMS API Client was designed using Python 3.10.
@@ -38,6 +38,25 @@ out_folder = "/home/myuser"
 dds_api.download_item(out_folder)
 ```
 
+### Search STAC Catalog
+
+The `Search_API` class provides helpers for searching the EODMS STAC catalog using [pystac-client](https://pystac-client.readthedocs.io/).
+
+```python
+from eodms import aaa, dds, search
+
+aaa_api = aaa.AAA_API('myeodmsusername', 'myeodmspassword')
+
+# Build an authenticated pystac-client
+client = search.build_stac_client(aaa_api=aaa_api, environment='prod')
+
+# List available collections
+for collection in client.get_collections():
+    print(collection.id)
+```
+
+For a full search-and-download example see [stac_dds_test.py](./tests/stac_dds_test.py) and [features_dds_test.py](./tests/features_dds_test.py).
+
 ## Testing
 
 ### Clone Repository
@@ -48,7 +67,7 @@ git clone https://github.com/eodms-sgdot/eodms-py.git
 
 ### Package Installation
 
-If the pip installation above did not work, you can install from the cloned repository:
+Install the package in editable mode so that changes to the source are reflected immediately:
 
 ```bash
 cd eodms-py
@@ -83,18 +102,80 @@ Options:
 python rapi_dds_test.py -u eodms_user -p eodms_pwd -c RCMImageProducts
 ```
 
+### Run stac_dds_test.py
+
+Searches the EODMS STAC catalog using the OGC Features items endpoint (paginated) and downloads the first result via DDS. Omitting `--collection` lists all available collections and their queryable fields.
+
+```
+Usage: stac_dds_test.py [OPTIONS]
+
+  Search and Download images from EODMS STAC catalog and DDS.
+
+Options:
+  -u, --username TEXT      The EODMS username.
+  -p, --password TEXT      The EODMS password.
+  -c, --collection TEXT    The collection name.
+      --uuid TEXT          UUID of the image to download (skips search).
+  -d, --datetime TEXT      Temporal filter as ISO 8601 string or range
+                           (e.g., "2023-01-01/2023-12-31").
+  -b, --bbox TEXT          Bounding box: west,south,east,north
+                           (e.g., "-100,45,-95,50").
+  -l, --limit INTEGER      Maximum number of items to fetch (default: 1000).
+  -f, --filter TEXT        CQL2 text filter expression
+                           (e.g., "roll_number = 'KA3'").
+      --output TEXT        Output GeoJSON filename (e.g., results.geojson).
+  -e, --env TEXT           Environment: "prod" (default) or "staging".
+  -o, --out_folder TEXT    Output folder (default: current folder).
+  -h, --help               Show this message and exit.
+```
+
+```bash
+# List all collections and queryable fields
+python stac_dds_test.py
+
+# Search and download first RCM image
+python stac_dds_test.py -u eodms_user -p eodms_pwd -c RCMImageProducts
+
+# Search with datetime and bbox filters
+python stac_dds_test.py -u eodms_user -p eodms_pwd -c RCMImageProducts -d "2023-01-01/2023-12-31" -b "-100,45,-95,50"
+
+# Apply a CQL2 filter and save results as GeoJSON
+python stac_dds_test.py -u eodms_user -p eodms_pwd -c RCMImageProducts -f "roll_number = 'KA3'" --output results.geojson
+```
+
+### Run features_dds_test.py
+
+Browses the EODMS OGC Features API directly (without pystac-client) with page-token pagination, and downloads a specific feature via DDS.
+
+```
+Usage: features_dds_test.py [OPTIONS]
+
+  OGC Features CLI for EODMS STAC
+
+Options:
+  -u, --username TEXT      The EODMS username.
+  -p, --password TEXT      The EODMS password.
+  -c, --collection TEXT    The collection name.
+  -f, --feature_id TEXT    The feature (item) ID.
+  -b, --bbox TEXT          Bounding box: west,south,east,north
+                           (e.g., "-100,45,-95,50").
+  -d, --datetime TEXT      Temporal filter as ISO 8601 string or range.
+  -e, --env TEXT           Environment: "prod" (default) or "staging".
+  -l, --limit INTEGER      Maximum number of features to return (default: 10).
+  -h, --help               Show this message and exit.
+```
+
+```bash
+# List all collections
+python features_dds_test.py
+
+# List features in a collection
+python features_dds_test.py -u eodms_user -p eodms_pwd -c RCMImageProducts -l 5
+
+# Download a specific feature by ID
+python features_dds_test.py -u eodms_user -p eodms_pwd -c RCMImageProducts -f some-feature-uuid
+```
+
 ## Documentation
 
 Official Swagger documentation can be found here, https://eodms-sgdot.nrcan-rncan.gc.ca/dds/v1/swagger-ui/#/
-
-## Cutover Plan
-
-Here is the plan to meet the cutover target of Mar 31 2026:
-
-1. Developers should import *both* `py-eodms-rapi` and `py-eodms-dds` into their code setup.
-2. For `rapi.search(..)` calls... you **don't** need to change these.
-3. For `rapi.order(..)` calls... you **do** need to change these and switch them over to `dds_api.get_item(..)` calls. See [rapi_dds_test.py](https://github.com/eodms-sgdot/py-eodms-dds/blob/main/tests/rapi_dds_test.py) an example.
-4. A new version of eodms-cli will also be released to handle the switch to py-eodms-dds. See: https://github.com/eodms-sgdot/eodms-cli/issues/46.
-5. Once all users are switched over, the https://www.eodms-sgdot.nrcan-rncan.gc.ca/wes/rapi/order endpoint will be blocked. Downloads will only be possible from the new endpoint, https://eodms-sgdot.nrcan-rncan.gc.ca/dds 
-
-The main difference is that `rapi.order` uses a `sequence_id`, whereas `get_item` only accepts a `uuid`. Thankfully, `uuid` is already returned by `rapi.search(..)`.
